+++ android/src/main/java/com/jhomlala/better_player/BetterPlayerPlugin.java
@@ -143,6 +143,7 @@ public class BetterPlayerPlugin implements FlutterPlugin, ActivityAware, MethodC
         if (flutterState == null) {
             Log.wtf(TAG, "Detached from the engine before registering to it.");
         }
+        disposeAllPlayers();
         BetterPlayerCache.releaseCache();
         flutterState.stopListening();
         flutterState = null;