+++ lib/src/configuration/better_player_event_type.dart
@@ -19,4 +19,5 @@ enum BetterPlayerEventType {
   changedResolution,
   pipStart,
   pipStop,
+  setupDataSource
 }