+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added initialStartIndex in BetterPlayerPlaylistConfiguration.
 * Fixed issue where player did not disposed properly on app quit.
 * Added placeholder parameter in BetterPlayerDataSource.
+* Fixed custom material full screen icons (by https://github.com/FelipeFernandesLeandro)
 
 ## 0.0.59
 * Fixed WEBVTT subtitles parsing.